Infobox added to plot
To make the plot more of a general readout of the run, I’ve added an
infobox that we can use to display, for example, eccentricities of the
orbits. Prettiness of plots has been improved over the last two commits
also. Note that as of commit #ad67604, hyperdrive now requires a
working LaTeX installation, mostly because matplotlib tables can go to
hell.